Introduce FixerFactory
Conflicts:
	Symfony/CS/Fixer.php
	Symfony/CS/Tests/FixerTest.php

Conflicts:
	Symfony/CS/Tests/FixerTest.php
